requirements.txt
python-telegram-bot==20.7
schedule==1.2.1
pytz==2023.3
aiohttp==3.9.1
nest-asyncio==1.5.8
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
ta==0.11.0  # Technical analysis library (alternativa ao TA-Lib)
psutil==7.0.0  # Para gerenciamento de processos e monitoramento de sistema
# TA-Lib==0.4.28  # Biblioteca de análise técnica mais completa - Requer instalação separada, veja o README.md 